+++ src/Core/planetClass.java
@@ -21,19 +21,15 @@ public class planetClass extends planetCore implements Serializable {
     /** planet construction methods **/
 
     //planet variables
-    private long distanceFromStar;
     private int planetRadius;
     private boolean isTidallyLocked;
     private int planetType;
-    //private boolean isInHabitableZone;
     private starClass parentStar;
     private boolean isPlanetHabitable;
     private boolean isHabited;
     private int planetNumber;
-    private int arrayLoc;
+    private transient int arrayLoc;
     private String planetName;
-    private String planetDesc;
-    private int numOfMoons;
     private colonyCore planetColony;
     private double resources;
     private int colonyID;
@@ -43,34 +39,32 @@ public class planetClass extends planetCore implements Serializable {
     public planetClass(starClass parentStar, int planetNumber) {
         this.parentStar = parentStar; //the parent star of the planet
         this.planetNumber = planetNumber; //the current planet being generated, with a higher number being further from the parent star
-        this.distanceFromStar = determineDistanceFromStar(this.parentStar, this.planetNumber);
+        //this.distanceFromStar = determineDistanceFromStar(this.parentStar, this.planetNumber);
         this.planetRadius = calculateSize(); //determines the radius of the planet in km
         this.isTidallyLocked = checkTidalLock(); //runs a check to determine tidal lock
-        //this.isInHabitableZone = isInHabitableZone(this.distanceFromStar, parentStar.getHabitableZoneMax(), parentStar.getHabitableZoneMin());
-        this.planetType = determinePlanetClass(this.distanceFromStar, this.planetRadius, this.isTidallyLocked, this.parentStar.isStarHabitable(), this.parentStar); //determines the planet's class
+        this.planetType = determinePlanetClass(0, this.planetRadius, this.isTidallyLocked, this.parentStar.isStarHabitable(), this.parentStar); //determines the planet's class
         this.arrayLoc = getPlanetFromID(planetType); //Gets the planet's array location from the ID.
         this.isPlanetHabitable = determineHabitability(arrayLoc); //whether or not the planet can be colonized
         this.isHabited = false; //by default, the planet is not currently colonized
         this.planetName = this.parentStar.getStarName() + " " + this.planetNumber;
+        this.resources = calculateResources(arrayLoc) * (gameSettings.currentResources / 100);
         //System.out.println("Planet '" + listOfPlanets.get(arrayLoc).getClassName() + "' (ID" + planetType + ") successfully generated.");
 
     }
 
     //generates a pre-defined planet
-    public planetClass(starClass parentStar, String planetName, int planetType, int planetNumber, long distanceFromStar, int planetRadius, boolean isInHabitableZone, boolean isHabited, boolean isPlanetHabitable, boolean isTidallyLocked, int numOfMoons, boolean homePlanet) {
+    public planetClass(starClass parentStar, String planetName, int planetType, int planetNumber, int planetRadius, boolean isHabited, boolean isPlanetHabitable, boolean isTidallyLocked, boolean homePlanet, double resources) {
         this.parentStar = parentStar;
         this.planetName = planetName;
         this.planetType = planetType;
         this.planetNumber = planetNumber;
-        this.distanceFromStar = distanceFromStar;
         this.planetRadius = planetRadius;
-        //this.isInHabitableZone = isInHabitableZone;
         this.arrayLoc = getPlanetFromID(this.planetType);
         this.isHabited = isHabited;
         this.isPlanetHabitable = isPlanetHabitable;
         this.isTidallyLocked = isTidallyLocked;
-        this.numOfMoons = numOfMoons;
         this.homePlanet = homePlanet;
+        this.resources = resources;
         //System.out.println("Predefined planet '" + listOfPlanets.get(arrayLoc).getClassName() + " [" + this.planetName + "] (ID" + planetType + ") successfully generated in the " + parentStar.getStarName() + " system.");
 
     }
@@ -79,14 +73,12 @@ public class planetClass extends planetCore implements Serializable {
 
     public boolean getHabited() { return this.isHabited; }
     public boolean getHabitability() { return this.isPlanetHabitable; }
-    public long getDistanceFromStar() { return this.distanceFromStar; }
     public int getPlanetRadius() { return this.planetRadius; }
     public String getPlanetName() { return this.planetName; }
     public boolean getTidalLock() { return this.isTidallyLocked; }
     public starClass getParentStar() { return this.parentStar; }
     public int getPlanetNumber() { return this.planetNumber; }
     public int getPlanetType() { return this.planetType; }
-    public int getNumOfMoons() { return this.numOfMoons; }
     public colonyCore getPlanetColony() { return this.planetColony; }
     public double getResources() { return this.resources; }
     public int getArrayLoc() { return this.arrayLoc; }
@@ -108,6 +100,9 @@ public class planetClass extends planetCore implements Serializable {
 
     public void setResources(double used) {
         this.resources = this.resources - used;
+        if (this.resources < 0) { //should go below zeor anyways, but redundancies
+            this.resources = 0;
+        }
     }
 
 }