+++ src/Core/turnTicker.java
@@ -41,9 +41,7 @@ public class turnTicker extends Thread {
         }
 
         gameSettings.currentDate++;
-        if (gameSettings.currentDate % 10 == 0) { //once every 10 turns increase the player's funds/resources/etc
-            gameSettings.player.turnTick(); //ticks the player data
-        }
+        gameSettings.player.tickStats(); //update player info
         gameSettings.ui.turnTick(); //refresh the UI
 
         runTurn();